Add files via upload
dataset.py:
File that contains the dataset used for the torch DataLoader

models.py:
File that contains the ResNet based upscaler and the DenseNet based upscaler, the discriminator model and the vgg19 based loss function

train_gan.py
Script which can be used to train the ResNet or DenseNet based upscaler. You can choose which one  to train by editing the section "# Initialize models
" starting at line 55. The DenseNet based network trains a bit faster, but the quality in comparison to the ResNet based generator has to be evaluated.